Movie 클래스의 rating 변수형을 double에서 float으로 수정 Movie 클래스 객체의 정보를 수정 및 조회할 수 있는 MovieItemRepository 클래스 생성 MovieController 클래스에 영화 목록을 조회할 수 있는 items 메소드 생성 MovieController 클래스에 영화 정보를 조회할 수 있는 item 메소드 생성

+++ Service_movie/src/main/java/com/example/MovieService/Movie.java
@@ -3,13 +3,12 @@ package com.example.MovieService;
 import lombok.Data;
 import java.util.Date;
 
-
 @Data
 public class Movie {
 
     private Long Id;            // 영화 고유 ID (서비스에 등록된 순서, 자동 증가)
     private String itemName;    // 영화 이름 or 영화 ID
-    private double rating;      // 영화 평점 ex) 0.0 ~ 5.0
+    private float rating;       // 영화 평점 ex) 0.0 ~ 5.0
     private long total;         // 누적 관객수
     private String genre;       // 영화 장르
     private Date releaseDate;   // 영화 개봉 일자
@@ -20,7 +19,7 @@ public class Movie {
     public Movie() {
     }
 
-    public Movie(String itemName, double rating, long total, String genre, Date releaseDate, String director, String actor, String info) {
+    public Movie(String itemName, float rating, long total, String genre, Date releaseDate, String director, String actor, String info) {
         this.itemName = itemName;
         this.rating = rating;
         this.total = total;
